requirement.txt
fastapi
uvicorn
gitpython
chromadb
google-generativeai
langchain
langchain-community
langchain-google-genai
tiktoken
